testprj/requirements.txt
Django
mysqlclient
djangorestframework
django-filter
django-cors-headers